requirements.txt
torch>=1.9
torchvision
timm
Pillow
numpy
tqdm
